Make something that works
So, it seems that I can't declare function pointers in macros, so we're working
with a fully-fledged function for Test now.

The workflow for setting up the tests is kinda ugly, and I need a better way to
track state of the tests, but this is a good start.  At least it offers me the
ability to have formatted output in my tests for debugging purposes.  That was
my number one goal for this little framework